requirements.txt
accelerate==1.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens 
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
boto3==1.36.4
botocore==1.36.4
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama 
coloredlogs==15.0.1
comm 
cpm-kernels==1.0.11
cryptography==44.0.0
dataclasses-json==0.6.7
debugpy
decorator 
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
durationpy==0.9
einops==0.8.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup 
executing 
faiss-cpu==1.9.0.post1
fastapi==0.115.7
fastcore==1.7.28
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.21
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel 
ipython 
ipython-genutils==0.2.0
ipywidgets==7.7.2
isoduration==20.11.0
jedi
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client 
jupyter_core 
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==1.1.11
kubernetes==31.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-huggingface==0.1.2
langchain-llm==0.4.15
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.3.1
loguru==0.7.3
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib-inline 
mdurl==0.1.2
mistune==3.1.0
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.23.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest_asyncio 
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.20.1
openai==1.60.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging
pandas==2.2.3
pandocfilters==1.5.1
parso 
peft==0.14.0
pickleshare 
pillow==11.1.0
platformdirs
posthog==3.9.2
prometheus_client==0.21.1
prompt_toolkit 
propcache==0.2.1
protobuf==5.29.3
psutil 
pure_eval 
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments 
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil 
python-dotenv==1.0.1
python-iso639==2025.1.27
python-json-logger==3.2.1
python-magic==0.4.27
python-oxmsg==0.0.1
pytz==2024.2
pywin32==307
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq 
RapidFuzz==3.11.0
rapidocr-onnxruntime==1.4.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
Send2Trash==1.8.3
sentence-transformers==3.3.1
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
six 
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack_data 
starlette==0.45.2
stqdm==0.0.5
streamlit==1.41.1
streamlit-jupyter==0.2.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado 
tqdm==4.67.1
traitlets 
transformers==4.48.1
transformers-stream-generator==0.0.5
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions 
tzdata==2025.1
unstructured==0.16.16
unstructured-client==0.29.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth 
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
widgetsnbextension==3.6.10
win32_setctime==1.2.0
wrapt==1.17.2
yarl==1.18.3
zipp
zstandard==0.23.0
